pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["core", "db", "connectors"]

[project]
name = "de-assist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "email-reply-parser>=0.5.12",
    "google-api-python-client>=2.124.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "openai>=1.14.0",
    "pgvector>=0.2.5",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0.1",
    "tiktoken>=0.6.0",
]

[project.scripts]
de-assist = "main:main"
telegram-bot = "connectors.telegram_bot:main"
gmail-ingest = "connectors.gmail_ingestion:ingest_recent_emails"

[tool.uv.sources]

[tool.pytest.ini_options]
pythonpath = ["."]
